Improve discoverability of preprocessing layers among keras layers
With the new symbol location of preprocessing layers directly in the keras layers
namespace, we should be more explicit in the doc strings that these layers
are preprocessing layers.

We should also link the preprocessing layer guide, which is the only convenient way
to find the full set of these layers on tensorflow.org.

PiperOrigin-RevId: 391873446